Move QBID and add references
