+++ openfisca_us/variables/irs/outputs.py
@@ -91,36 +91,6 @@ class was_plus_sey(Variable):
         )
 
 
-class rptc(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Refundable Payroll Tax Credit"
-    documentation = "Refundable Payroll Tax Credit for filing unit"
-    unit = USD
-
-
-refundable_payroll_tax_credit = variable_alias(
-    "refundable_payroll_tax_credit", rptc
-)
-
-
-class rptc_p(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    documentation = "Refundable Payroll Tax Credit for taxpayer"
-    unit = USD
-
-
-class rptc_s(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    documentation = "Refundable Payroll Tax Credit for spouse"
-    unit = USD
-
-
 class exact(Variable):
     value_type = bool
     entity = TaxUnit
@@ -170,14 +140,6 @@ class expanded_income(Variable):
         )
 
 
-class num(Variable):
-    value_type = int
-    entity = TaxUnit
-    definition_period = YEAR
-    documentation = "2 when MARS is 2 (married filing jointly); otherwise 1"
-    unit = USD
-
-
 class othertaxes(Variable):
     value_type = float
     entity = TaxUnit
@@ -213,153 +175,6 @@ class surtax(Variable):
         "search taxcalc/calcfunctions.py for how calculated and used"
     )
     unit = USD
-
-
-class taxbc(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Regular tax before credits"
-    documentation = "Regular tax on regular taxable income before credits"
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        capital_gains = parameters(period).irs.capital_gains.brackets
-        mars = tax_unit("mars", period)
-        dwks1 = tax_unit("c04800", period)
-
-        dwks16 = min_(capital_gains.thresholds["1"][mars], dwks1)
-        dwks17 = min_(tax_unit("dwks14", period), dwks16)
-
-        dwks20 = dwks16 - dwks17
-        lowest_rate_tax = capital_gains.rates["1"] * dwks20
-        # Break in worksheet lines
-        dwks13 = tax_unit("dwks13", period)
-        dwks21 = min_(dwks1, dwks13)
-        dwks22 = dwks20
-        dwks23 = max_(0, dwks21 - dwks22)
-        dwks25 = min_(capital_gains.thresholds["2"][mars], dwks1)
-        dwks19 = tax_unit("dwks19", period)
-        dwks26 = min_(dwks19, dwks20)
-        dwks27 = max_(0, dwks25 - dwks26)
-        dwks28 = min_(dwks23, dwks27)
-        dwks29 = capital_gains.rates["2"] * dwks28
-        dwks30 = dwks22 + dwks28
-        dwks31 = dwks21 - dwks30
-        dwks32 = capital_gains.rates["3"] * dwks31
-        # Break in worksheet lines
-        dwks33 = min_(
-            tax_unit("dwks9", period), tax_unit("filer_e24515", period)
-        )
-        dwks10 = tax_unit("dwks10", period)
-        dwks34 = dwks10 + dwks19
-        dwks36 = max_(0, dwks34 - dwks1)
-        dwks37 = max_(0, dwks33 - dwks36)
-        dwks38 = 0.25 * dwks37
-        # Break in worksheet lines
-        dwks39 = dwks19 + dwks20 + dwks28 + dwks31 + dwks37
-        dwks40 = dwks1 - dwks39
-        dwks41 = 0.28 * dwks40
-
-        # SchXYZ call in Tax-Calculator
-
-        # Separate non-negative taxable income into two non-negative components,
-        # doing this in a way so that the components add up to taxable income
-        # define pass-through income eligible for PT schedule
-        individual_income = parameters(period).irs.income
-        e26270 = tax_unit("filer_e26270", period)
-        e00900 = tax_unit("filer_e00900", period)
-
-        # Determine pass-through and non-pass-through income
-        pt_active_gross = e00900 + e26270
-        pt_active = pt_active_gross
-        pt_active = min_(pt_active, e00900 + e26270)
-        pt_taxinc = max_(0, pt_active)
-        taxable_income = dwks19
-
-        pt_taxinc = min_(pt_taxinc, taxable_income)
-        reg_taxinc = max_(0, taxable_income - pt_taxinc)
-        pt_tbase = reg_taxinc
-
-        mars = tax_unit("mars", period)
-
-        # Initialise regular and pass-through income tax to zero
-        reg_tax = 0
-        pt_tax = 0
-        last_reg_threshold = 0
-        last_pt_threshold = 0
-        for i in range(1, 7):
-            # Calculate rate applied to regular income up to the current
-            # threshold (on income above the last threshold)
-            reg_threshold = individual_income.bracket.thresholds[str(i)][mars]
-            amount_in_bracket = amount_between(
-                reg_taxinc, last_reg_threshold, reg_threshold
-            )
-            reg_tax += (
-                individual_income.bracket.rates[str(i)] * amount_in_bracket
-            )
-            last_reg_threshold = reg_threshold
-
-            # Calculate rate applied to pass-through income on in the same
-            # way, but as treated as if stacked on top of regular income
-            # (which is not taxed again)
-            pt_threshold = max_(
-                individual_income.pass_through.bracket.thresholds[str(i)][mars]
-                - pt_tbase,
-                0,
-            )
-            pt_tax += individual_income.pass_through.bracket.rates[
-                str(i)
-            ] * amount_between(pt_taxinc, last_pt_threshold, pt_threshold)
-            last_pt_threshold = pt_threshold
-
-        # Calculate regular and pass-through tax above the last threshold
-        reg_tax += individual_income.bracket.rates["7"] * max_(
-            reg_taxinc - last_reg_threshold, 0
-        )
-        pt_tax += individual_income.pass_through.bracket.rates["7"] * max_(
-            pt_taxinc - last_pt_threshold, 0
-        )
-
-        dwks42 = reg_tax + pt_tax
-
-        dwks43 = sum(
-            [
-                dwks29,
-                dwks32,
-                dwks38,
-                dwks41,
-                dwks42,
-                lowest_rate_tax,
-            ]
-        )
-        c05200 = tax_unit("income_tax_main_rates", period)
-        dwks44 = c05200
-        dwks45 = min_(dwks43, dwks44)
-
-        hasqdivltcg = tax_unit("hasqdivltcg", period)
-
-        return where(hasqdivltcg, dwks45, c05200)
-
-
-regular_tax_before_credits = variable_alias(
-    "regular_tax_before_credits", taxbc
-)
-
-
-class c00100(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "AGI"
-    documentation = "Adjusted Gross Income"
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        return add(tax_unit, period, ["ymod1", "c02500", "c02900"])
-
-
-adjusted_gross_income = variable_alias("adjusted_gross_income", c00100)
 
 
 class c01000(Variable):
@@ -514,168 +329,6 @@ class c04600(Variable):
         line_6 = line_5 / (2500 / tax_unit("sep", period))
         line_7 = phaseout.rate * line_6
         return tax_unit("pre_c04600", period) * (1 - line_7)
-
-
-class qbided(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "QBI deduction"
-    documentation = "Qualified Business Income (QBI) deduction"
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        mars = tax_unit("mars", period)
-        QBINC_ELEMENTS = [
-            "filer_e00900",
-            "filer_e26270",
-            "filer_e02100",
-            "filer_e27200",
-        ]
-        qbinc = max_(0, add(tax_unit, period, QBINC_ELEMENTS))
-        qbid = parameters(period).irs.deductions.qualified_business_interest
-        lower_threshold = qbid.threshold.lower[mars]
-        upper_threshold = lower_threshold + qbid.threshold.gap[mars]
-        pre_qbid_taxinc = tax_unit("pre_qbid_taxinc", period)
-        under_lower_threshold = pre_qbid_taxinc < lower_threshold
-        between_thresholds = ~under_lower_threshold & (
-            pre_qbid_taxinc < upper_threshold
-        )
-        above_upper_threshold = ~under_lower_threshold & ~between_thresholds
-        income_is_qualified = tax_unit("pt_sstb_income", period)
-
-        # Wage/capital limitations
-        w2_wages = tax_unit("pt_binc_w2_wages", period)
-        business_property = tax_unit("pt_ubia_property", period)
-        wage_cap = w2_wages * qbid.cap.w2_wages.rate
-        alt_cap = (
-            w2_wages * qbid.cap.w2_wages.alt_rate
-            + business_property * qbid.cap.business_property.rate
-        )
-        fraction_of_gap_passed = (
-            pre_qbid_taxinc - lower_threshold
-        ) / qbid.threshold.gap[mars]
-        fraction_of_gap_unused = (
-            upper_threshold - pre_qbid_taxinc
-        ) / qbid.threshold.gap[mars]
-
-        # Adjustments for qualified income under the upper threshold
-        qbi_between_threshold_multiplier = where(
-            income_is_qualified & between_thresholds,
-            fraction_of_gap_unused,
-            1.0,
-        )
-        max_qbid = (
-            qbinc * qbid.pass_through_rate * qbi_between_threshold_multiplier
-        )
-        full_cap = max_(wage_cap, alt_cap) * qbi_between_threshold_multiplier
-
-        # Adjustment for QBID where income is between the main thresholds
-        adjustment = fraction_of_gap_passed * (max_qbid - full_cap)
-
-        qbid_amount = select(
-            (
-                under_lower_threshold,
-                between_thresholds,
-                above_upper_threshold,
-            ),
-            (
-                max_qbid,
-                max_qbid - adjustment,
-                where(income_is_qualified, 0, min_(max_qbid, full_cap)),
-            ),
-        )
-
-        # Apply taxable income cap
-        net_cg = add(tax_unit, period, ["filer_e00650", "c01000"])
-        taxinc_cap = qbid.pass_through_rate * max_(0, pre_qbid_taxinc - net_cg)
-        return min_(qbid_amount, taxinc_cap)
-
-
-class c04800(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Taxable income"
-    documentation = "Regular taxable income"
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        return max_(
-            0, tax_unit("pre_qbid_taxinc", period) - tax_unit("qbided", period)
-        )
-
-
-taxable_income = variable_alias("taxable_income", c04800)
-
-
-class c05200(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Income tax main rates"
-    unit = USD
-    documentation = "Tax amount from Sch X,Y,Z tables"
-
-    def formula(tax_unit, period, parameters):
-        # Separate non-negative taxable income into two non-negative components,
-        # doing this in a way so that the components add up to taxable income
-        # define pass-through income eligible for PT schedule
-        individual_income = parameters(period).irs.income
-        e26270 = tax_unit("filer_e26270", period)
-        e00900 = tax_unit("filer_e00900", period)
-
-        # Determine pass-through and non-pass-through income
-        pt_active_gross = e00900 + e26270
-        pt_active = pt_active_gross
-        pt_active = min_(pt_active, e00900 + e26270)
-        pt_taxinc = max_(0, pt_active)
-        taxable_income = tax_unit("c04800", period)
-
-        pt_taxinc = min_(pt_taxinc, taxable_income)
-        reg_taxinc = max_(0, taxable_income - pt_taxinc)
-        pt_tbase = reg_taxinc
-
-        mars = tax_unit("mars", period)
-
-        # Initialise regular and pass-through income tax to zero
-        reg_tax = 0
-        pt_tax = 0
-        last_reg_threshold = 0
-        last_pt_threshold = 0
-        for i in range(1, 7):
-            # Calculate rate applied to regular income up to the current
-            # threshold (on income above the last threshold)
-            reg_threshold = individual_income.bracket.thresholds[str(i)][mars]
-            reg_tax += individual_income.bracket.rates[
-                str(i)
-            ] * amount_between(reg_taxinc, last_reg_threshold, reg_threshold)
-            last_reg_threshold = reg_threshold
-
-            # Calculate rate applied to pass-through income on in the same
-            # way, but as treated as if stacked on top of regular income
-            # (which is not taxed again)
-            pt_threshold = max_(
-                0,
-                individual_income.pass_through.bracket.thresholds[str(i)][mars]
-                - pt_tbase,
-            )
-            pt_tax += individual_income.pass_through.bracket.rates[
-                str(i)
-            ] * amount_between(pt_taxinc, last_pt_threshold, pt_threshold)
-            last_pt_threshold = pt_threshold
-
-        # Calculate regular and pass-through tax above the last threshold
-        reg_tax += individual_income.bracket.rates["7"] * max_(
-            reg_taxinc - last_reg_threshold, 0
-        )
-        pt_tax += individual_income.pass_through.bracket.rates["7"] * max_(
-            pt_taxinc - last_pt_threshold, 0
-        )
-        return reg_tax + pt_tax
-
-
-income_tax_main_rates = variable_alias("income_tax_main_rates", c05200)
 
 
 class c05700(Variable):
